Revert "fix: rename ui-> prebuiltui to avoid import conflicts"
This reverts commit c654d68045d1d25579a55cc413f2383920b2fd29.
